telegram/requirements.txt
paramiko==3.3.1
python-telegram-bot==20.6
PyYAML==6.0.1
requests==2.31.0
